refactor: extract settings page container
