+++ src/pages/settings/ui/SettingsPage.spec.tsx
@@ -1,4 +1,4 @@
-import { fireEvent, render, screen, waitFor } from "@testing-library/react";
+import { fireEvent, render, screen } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import "@testing-library/jest-dom/vitest";
 
@@ -45,24 +45,5 @@ describe("SettingsPage", () => {
     fireEvent.keyDown(window, { key: "t" });
 
     expect(mocks.navigate).toHaveBeenCalledExactlyOnceWith("/");
-  });
-
-  it("displays an alert when sign-out fails and allows retrying sign-out", async () => {
-    mocks.authAccount = { uid: "retry-uid-a", displayName: "Test User" };
-    mocks.authUid = "retry-uid-a";
-    const signOutError = new Error("sign out failed");
-    const logout = vi.fn().mockRejectedValueOnce(signOutError).mockResolvedValueOnce(undefined);
-
-    render(<SettingsPage login={vi.fn()} logout={logout} />);
-
-    fireEvent.click(screen.getByRole("button", { name: "Logout" }));
-
-    expect(await screen.findByRole("alert")).toHaveTextContent("Unable to sign out.");
-    expect(logout).toHaveBeenCalledOnce();
-
-    fireEvent.click(screen.getByRole("button", { name: "Retry" }));
-
-    await waitFor(() => expect(screen.queryByRole("alert")).not.toBeInTheDocument());
-    expect(logout).toHaveBeenCalledTimes(2);
   });
 });
